Add files via upload
Hey professor, I got together with Mike Bartoli and Mike Durso. We did the coding separately from one another and only conversed ideas and did not copy each others code. We all came to the same problem when trying to evaluate the temperature and the relative humidity. We do not get the correct answers. We get different answers from one another but the answers are still not correct. We have tried everything we could think of and yet we still ran into the same problem.